+++ NINATest/Sequencer/SequenceItem/Dome/CloseDomeShutterTest.cs
@@ -77,7 +77,7 @@ namespace NINATest.Sequencer.SequenceItem.Dome {
 
         [Test]
         public async Task Execute_NoIssues_LogicCalled() {
-            domeMediatorMock.Setup(x => x.GetInfo()).Returns(new DomeInfo() { Connected = true });
+            domeMediatorMock.Setup(x => x.GetInfo()).Returns(new DomeInfo() { Connected = true, ShutterStatus = NINA.Equipment.Interfaces.ShutterState.ShutterClosed });
             domeMediatorMock.Setup(x => x.CloseShutter(It.IsAny<CancellationToken>())).Returns(Task.FromResult(true));
 
             var sut = new CloseDomeShutter(domeMediatorMock.Object);