+++ NINATest/Sequencer/SequenceItem/FlatDevice/OpenCoverTest.cs
@@ -81,7 +81,7 @@ namespace NINATest.Sequencer.SequenceItem.FlatDevice {
 
         [Test]
         public async Task Execute_NoIssues_LogicCalled() {
-            fdMediatorMock.Setup(x => x.GetInfo()).Returns(new FlatDeviceInfo() { Connected = true, SupportsOpenClose = true });
+            fdMediatorMock.Setup(x => x.GetInfo()).Returns(new FlatDeviceInfo() { Connected = true, SupportsOpenClose = true, CoverState = NINA.Equipment.Interfaces.CoverState.Open });
 
             var sut = new OpenCover(fdMediatorMock.Object);
             await sut.Execute(default, default);